a6000: overlay: Kang autobrightness curve from Xiaomi Redmi 3S
Based on recent logcats by a daily user
- The user claimed to have been using this rom from past 3 months and have transitted
from most of the updates he used a lux meter or some other metering equipment to
properly meter the brightness curve of Lenovo A6000 for different light
conditions and according to his data yes i was using very inaccurate values
and the second part was that the user isn't properly able to tune the brightness
to his choice so add forbidden restrictions for brigtness tunables and let
the users go all the way down the brightness curve .

This fixes the auto brightness curve being somewhat dark. The Xiaomi Redmi 3S has
the same panel as we do and its curve was gained by using proper
metering equipment, so just use that one.

These values are extracted straight from the apk without modifications

* A revert worthy commit if not properly working

Signed-off-by: Hassan Sardar <hassansardar08@live.com>